refactor(dhcp): rename layer to dhcpv4
