+++ crafter/tests/fixtures/summaries/ipv4-udp-dhcp-discover.summary.txt
@@ -1,1 +1,1 @@
-Ipv4(src=0.0.0.0, dst=255.255.255.255, proto=udp(17)) / Udp(sport=68, dport=67, len=267, checksum_status=Valid) / Dhcp(type=discover, xid=0x3903f326, yiaddr=0.0.0.0)
+Ipv4(src=0.0.0.0, dst=255.255.255.255, proto=udp(17)) / Udp(sport=68, dport=67, len=267, checksum_status=Valid) / Dhcpv4(type=discover, xid=0x3903f326, yiaddr=0.0.0.0)
